pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexis"
version = "0.1.0"
description = "Hexis: Self-Learning Computer Use Agent with frozen VLM backbone + trainable expert heads"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "transformers>=4.45",
    "numpy>=1.26",
    "Pillow>=10.0",
    "playwright>=1.40",
    "anthropic>=0.39",
]

[project.scripts]
hexis = "hexis.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["hexis*"]
